pyproject.toml
[project]
name = "keine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-frontmatter>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
